requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
adversarial-robustness-toolbox=1.7.2=pypi_0
apex=0.1=pypi_0
argparse=1.4.0=pypi_0
blas=1.0=mkl
blosc=1.21.0=h8c45485_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py38h06a4308_0
coverage=5.5=py38h27cfd23_2
cox=0.1.post3=pyh9f0ad1d_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py38_0
cython=0.29.24=py38h295c915_0
dbus=1.13.18=hb2f20db_0
dill=0.3.4=pyhd3eb1b0_0
easydict=1.9=pypi_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.69.1=h5202010_0
gputil=1.4.0=pyh9f0ad1d_0
grpcio=1.36.1=py38h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=he6710b0_3
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py38h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.36.0=pypi_0
lz4-c=1.9.3=h295c915_1
lzo=2.10=h7b6447c_2
matplotlib=3.4.2=py38h06a4308_0
matplotlib-base=3.4.2=py38hab158f2_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.2=py38h51133e4_0
mock=4.0.3=pyhd3eb1b0_0
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
networkx=2.6.2=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
numba=0.53.1=pypi_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.21.2=pypi_0
olefile=0.46=pyhd3eb1b0_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.2=py38h8c16a72_0
patsy=0.5.2=pypi_0
pcre=8.45=h295c915_0
pillow=8.3.1=py38h2c7a002_0
pip=21.2.2=py38h06a4308_0
protobuf=3.17.2=py38h295c915_0
psutil=5.8.0=py38h27cfd23_1
py3nvml=0.2.5=py38h32f6830_2
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pytables=3.6.1=py38h9fd0a39_0
python=3.8.11=h12debd9_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scikit-learn=0.24.2=py38ha9443f7_0
scipy=1.7.1=py38h292c36d_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
statsmodels=0.13.0=pypi_0
tensorboardx=2.2=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=hbc83047_0
torchsummary=1.5.1=pypi_0
torchvision=0.8.2=py38_cu110
tornado=6.1=py38h27cfd23_0
tqdm=4.62.2=pyhd3eb1b0_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
xmltodict=0.12.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
